pyproject.toml
[project]
name = "ecg-explain"
version = "0.1.0"
description = "12-lead ECG classification with clinical interpretability"
authors = [{name = "Muhammed"}]
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
    "torch>=2.2",
    "numpy>=1.26",
    "pandas>=2.2",
    "scikit-learn>=1.4",
    "wfdb>=4.1",
    "scipy>=1.12",
    "matplotlib>=3.8",
    "tqdm>=4.66",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "ruff>=0.4", "mypy>=1.8", "pre-commit>=3.7"]
app = ["gradio>=4.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["src/ecg_explain"]

[tool.uv]
package = true

[tool.uv.sources]
ecg-explain = { workspace = true }

[tool.uv.workspace]
members = ["."]

[tool.mypy]
python_version = "3.11"
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
no_implicit_optional = true
check_untyped_defs = true
ignore_missing_imports = true   # third-party libs without stubs (wfdb, etc.)
files = ["src/ecg_explain"]

[dependency-groups]
dev = [
    "types-pyyaml>=6.0.12.20260408",
]
